requirements.txt
ipython
notebook
nb_conda_kernels

polars
python-dotenv
neo4j-python-driver

langchain
langchain-core
langchain-community
langchain-google-genai
